Chatbot_Structure/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
cryptography==41.0.3
openai==0.27.8
mysql-connector-python==8.0.33
passlib==1.7.4